Rebase, refactor useResourceRouteResolver composable file structure, set tile Imagedimension to available size
